users: add business logic

+++ src/users/service.ts
@@ -1,4 +1,8 @@
-import { InvalidArgumentError } from "../common/service_errors";
+import { Issuer } from "../auth/model";
+import {
+  InvalidArgumentError,
+  UnauthorizedError,
+} from "../common/service_errors";
 import { User, PartialUser, createUserSchema, updateUserSchema } from "./model";
 import Repository from "./repository";
 
@@ -18,7 +22,11 @@ async function createOne(user: User): Promise<User> {
   return { ...newUser, password: "[redacted]" };
 }
 
-async function getOne(id: Number): Promise<User | null> {
+async function getOne(id: Number, issuer: Issuer): Promise<User | null> {
+  if (["customer", "owner"].includes(issuer.role) && issuer.id != id) {
+    throw new UnauthorizedError("You can only see your own account.");
+  }
+
   const user = await Repository.getOne(id);
   if (user) {
     return { ...user, password: "[redacted]" };
@@ -32,7 +40,30 @@ async function getAll(): Promise<Array<User>> {
   });
 }
 
-async function updateOne(id: Number, user: PartialUser): Promise<User | null> {
+async function updateOne(
+  id: Number,
+  user: PartialUser,
+  issuer: Issuer,
+): Promise<User | null> {
+  if (["customer", "owner"].includes(issuer.role) && issuer.id != id) {
+    throw new UnauthorizedError("You can only update your own account.");
+  }
+
+  if (["customer", "owner"].includes(issuer.role) && user.role) {
+    throw new UnauthorizedError("You cannot change your role.");
+  }
+
+  if (issuer.role === "staff" && user.role === "admin") {
+    throw new UnauthorizedError("Only admins can create admins.");
+  }
+
+  if (
+    issuer.role === "staff" &&
+    (await Repository.getOne(id))?.role === "admin"
+  ) {
+    throw new UnauthorizedError("Staff cannot downgrade admins.");
+  }
+
   const { value, error } = updateUserSchema.validate(user);
   if (error) {
     throw error;
@@ -45,7 +76,15 @@ async function updateOne(id: Number, user: PartialUser): Promise<User | null> {
   return await Repository.updateOne(id, value);
 }
 
-async function deleteOne(id: Number): Promise<Number | null> {
+async function deleteOne(id: Number, issuer: Issuer): Promise<Number | null> {
+  if (["customer", "owner"].includes(issuer.role) && issuer.id != id) {
+    throw new UnauthorizedError("You can only delete your own account.");
+  }
+
+  if ((await Repository.getOne(id))?.role == "admin") {
+    throw new UnauthorizedError("You cannot delete an admin's account.");
+  }
+
   return await Repository.deleteOne(id);
 }
